Add files via upload
This is a C++ console-based Bookstore Management System that allows users and admins to interact with a book inventory. The project includes functionalities such as adding, updating, deleting, and viewing books, as well as cart, wishlist, payment, and order cancellation options. All data is managed using file handling with books.txt.